+++ lib/features/record/components/announcement_bar/notification_bar.dart
@@ -6,14 +6,14 @@ import 'package:pilll/provider/pill_sheet_group.dart';
 import 'package:pilll/provider/pilll_ads.dart';
 import 'package:pilll/features/premium_introduction/premium_introduction_sheet.dart';
 import 'package:pilll/features/premium_introduction/util/discount_deadline.dart';
-import 'package:pilll/features/record/components/notification_bar/components/discount_price_deadline.dart';
-import 'package:pilll/features/record/components/notification_bar/components/ended_pill_sheet.dart';
-import 'package:pilll/features/record/components/notification_bar/components/pilll_ads.dart';
-import 'package:pilll/features/record/components/notification_bar/components/user_survey.dart';
-import 'package:pilll/features/record/components/notification_bar/components/premium_trial_limit.dart';
-import 'package:pilll/features/record/components/notification_bar/components/recommend_signup.dart';
-import 'package:pilll/features/record/components/notification_bar/components/recommend_signup_premium.dart';
-import 'package:pilll/features/record/components/notification_bar/components/rest_duration.dart';
+import 'package:pilll/features/record/components/announcement_bar/components/discount_price_deadline.dart';
+import 'package:pilll/features/record/components/announcement_bar/components/ended_pill_sheet.dart';
+import 'package:pilll/features/record/components/announcement_bar/components/pilll_ads.dart';
+import 'package:pilll/features/record/components/announcement_bar/components/user_survey.dart';
+import 'package:pilll/features/record/components/announcement_bar/components/premium_trial_limit.dart';
+import 'package:pilll/features/record/components/announcement_bar/components/recommend_signup.dart';
+import 'package:pilll/features/record/components/announcement_bar/components/recommend_signup_premium.dart';
+import 'package:pilll/features/record/components/announcement_bar/components/rest_duration.dart';
 import 'package:pilll/features/sign_in/sign_in_sheet.dart';
 import 'package:pilll/provider/locale.dart';
 import 'package:pilll/provider/premium_and_trial.codegen.dart';
