+++ test/features/record/announcement_bar/announcement_bar_test.dart
@@ -4,14 +4,14 @@ import 'package:pilll/utils/analytics.dart';
 import 'package:pilll/provider/pill_sheet_group.dart';
 import 'package:pilll/provider/pilll_ads.dart';
 import 'package:pilll/features/premium_introduction/util/discount_deadline.dart';
-import 'package:pilll/features/record/components/notification_bar/components/discount_price_deadline.dart';
-import 'package:pilll/features/record/components/notification_bar/components/ended_pill_sheet.dart';
-import 'package:pilll/features/record/components/notification_bar/components/pilll_ads.dart';
-import 'package:pilll/features/record/components/notification_bar/notification_bar.dart';
-import 'package:pilll/features/record/components/notification_bar/components/premium_trial_limit.dart';
-import 'package:pilll/features/record/components/notification_bar/components/recommend_signup.dart';
-import 'package:pilll/features/record/components/notification_bar/components/recommend_signup_premium.dart';
-import 'package:pilll/features/record/components/notification_bar/components/rest_duration.dart';
+import 'package:pilll/features/record/components/announcement_bar/components/discount_price_deadline.dart';
+import 'package:pilll/features/record/components/announcement_bar/components/ended_pill_sheet.dart';
+import 'package:pilll/features/record/components/announcement_bar/components/pilll_ads.dart';
+import 'package:pilll/features/record/components/announcement_bar/notification_bar.dart';
+import 'package:pilll/features/record/components/announcement_bar/components/premium_trial_limit.dart';
+import 'package:pilll/features/record/components/announcement_bar/components/recommend_signup.dart';
+import 'package:pilll/features/record/components/announcement_bar/components/recommend_signup_premium.dart';
+import 'package:pilll/features/record/components/announcement_bar/components/rest_duration.dart';
 import 'package:pilll/entity/pill_sheet.codegen.dart';
 import 'package:pilll/entity/pill_sheet_group.codegen.dart';
 import 'package:pilll/entity/pill_sheet_type.dart';
